refactor(#792): split local_dispatcher_tests.rs 7-way (3404->575 + 6 siblings)
